backend/scripts/requirements.txt
requests==2.31.0
python-socketio==5.11.1
python-engineio==4.9.0
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
langchain==0.1.12
langchain-community==0.0.27
langchain-qdrant==0.0.5
qdrant-client==1.7.3
python-dotenv==1.0.1 